add a function to zdns to get local IP for v4 and v6
