+++ examples/multi_thread_lookup/multi_threaded.go
@@ -21,7 +21,6 @@ import (
 	"github.com/miekg/dns"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/zmap/zdns/v2/examples/utils"
 	"github.com/zmap/zdns/v2/src/zdns"
 )
 
@@ -71,7 +70,7 @@ func main() {
 func initializeResolver(cache *zdns.Cache) *zdns.Resolver {
 	// Create a ResolverConfig object
 	resolverConfig := zdns.NewResolverConfig()
-	localAddr, err := utils.GetLocalIPByConnecting()
+	localAddr, err := zdns.GetLocalIPv4Address()
 	if err != nil {
 		log.Fatal("Error getting local IP: ", err)
 	}
