--- examples/utils/utils.go
@@ -1,42 +1,0 @@
-/* ZDNS Copyright 2024 Regents of the University of Michigan
-*
-* Licensed under the Apache License, Version 2.0 (the "License"); you may not
-* use this file except in compliance with the License. You may obtain a copy
-* of the License at http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-* implied. See the License for the specific language governing
-* permissions and limitations under the License.
- */
-
-package utils
-
-import (
-	"net"
-
-	"github.com/pkg/errors"
-
-	log "github.com/sirupsen/logrus"
-)
-
-func GetLocalIPByConnecting() (net.IP, error) {
-	conn, err := net.Dial("udp", "1.1.1.1:53")
-	if err != nil {
-		return nil, err
-	}
-	defer func(conn net.Conn) {
-		err := conn.Close()
-		if err != nil {
-			log.Errorf("failed to close connection: %v", err)
-		}
-	}(conn)
-
-	localAddr, ok := conn.LocalAddr().(*net.UDPAddr)
-	if !ok {
-		return nil, errors.New("failed to get the local address as a UDP address")
-	}
-
-	return localAddr.IP, nil
-}
